+++ docs/changelog.rst
@@ -20,6 +20,7 @@ Changelog
 - 移除 ``xpaw.FormData`` , ``xpaw.URL``
 - 移除 ``xpaw.MultiDict`` , ``xpaw.CIMultiDict`` , 改由 ``xpaw.HttpHeaders`` 替代承载headers的功能
 - 为了和Python内置错误区分开，请求超时错误TimeoutError更名为RequestTimeout
+- ``default_headers`` 默认为 ``None`` , 浏览器默认的HTTP header改由UserAgentMiddleware根据设定的浏览器类型进行设置
 
 0.10.3 (2018-09-01)
 -------------------